+++ be/src/io/file_factory.cpp
@@ -51,6 +51,7 @@
 #include "util/uid_util.h"
 
 namespace doris {
+#include "common/compile_check_begin.h"
 
 constexpr std::string_view RANDOM_CACHE_BASE_PATH = "random";
 
@@ -87,7 +88,7 @@ int32_t get_broker_index(const std::vector<TNetworkAddress>& brokers, const std:
     }
 
     // secondly select broker by hash of file path
-    auto key = HashUtil::hash(path.data(), path.size(), 0);
+    auto key = HashUtil::hash(path.data(), static_cast<int32_t>(path.size()), 0);
     return key % brokers.size();
 }
 
@@ -274,5 +275,6 @@ Status FileFactory::create_pipe_reader(const TUniqueId& load_id, io::FileReaderS
 
     return Status::OK();
 }
+#include "common/compile_check_end.h"
 
 } // namespace doris
